Fix artificial work bug
Previously the artificial work loop was sharing a global variable among
all locales leading to contention. We want the artificial work
computation to be completely independent across all locales. This PR
rewrites the artificial work loop as a C extern with a variable that
will be local to each process.

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>